Isolated ResidueDepth and MSMS tests
